move pk downloader into own module
