+++ Cargo.toml
@@ -10,7 +10,9 @@ authors = [ "Finn Behrens <me@kloenk.dev>" ]
 default = [ "uf2" ]
 simulator = [ ]
 invert = [ ]
-updater = [ "alloc", "dep:pkrs", "dep:reqwest" ]
+downloaders = [ "alloc" ]
+downloader = [ "downloader-pk" ]
+downloader-pk = [ "downloaders", "dep:pkrs", "dep:reqwest" ]
 wasm = [ "dep:wasm-bindgen" ]
 defmt = [ "dep:defmt" ]
 uf2 = [ ]
