+++ cli/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-sysbadge = { path = "..", features = [ "uf2", "updater" ] }
+sysbadge = { path = "..", features = [ "uf2", "downloader" ] }
 
 tokio = { version = "1.32", features = [ "rt", "macros", "rt-multi-thread" ] }
 clap = { version = "4.3", features = [ "cargo", "derive" ] }
