+++ web/Cargo.toml
@@ -12,7 +12,7 @@ crate-type = ["cdylib"]
 
 [features]
 default = [ "badge", "update" ]
-update = [ "reqwest", "pkrs", "wasm-bindgen-futures", "sysbadge/updater", "sysbadge/uf2", "wasm-bindgen/serde-serialize", "web-sys/HtmlInputElement", "web-sys/HtmlButtonElement", "web-sys/Blob", "web-sys/Url", "web-sys/BlobPropertyBag" ]
+update = [ "reqwest", "pkrs", "wasm-bindgen-futures", "sysbadge/downloader", "sysbadge/uf2", "wasm-bindgen/serde-serialize", "web-sys/HtmlInputElement", "web-sys/HtmlButtonElement", "web-sys/Blob", "web-sys/Url", "web-sys/BlobPropertyBag" ]
 badge = [ "embedded-graphics", "embedded-graphics-web-simulator", "web-sys/HtmlCanvasElement", "web-sys/CanvasRenderingContext2d" ]
 
 [dependencies]
